fix(wmi): cast $3::int in Claire answer update (null param type inference)
Mid-round score is NULL; Postgres could not infer the bare param's type inside
the CASE, throwing 42P08. Explicit ::int cast resolves it.

+++ api/services/wmi/claire.ts
@@ -229,8 +229,8 @@ export async function answerClaireRound(
     await client.query(
       `UPDATE claire_drill_rounds
           SET responses = $2::jsonb,
-              score = $3,
-              completed_at = CASE WHEN $3 IS NOT NULL THEN COALESCE(completed_at, NOW()) ELSE completed_at END
+              score = $3::int,
+              completed_at = CASE WHEN $3::int IS NOT NULL THEN COALESCE(completed_at, NOW()) ELSE completed_at END
         WHERE id = $1`,
       [roundId, JSON.stringify(responses), score],
     )
